blog-automation/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.1
pytrends>=4.9
yfinance>=0.2.36
google-genai>=1.0
gspread>=6.0
google-auth>=2.27
python-dotenv>=1.0
tenacity>=8.2
